Added UserDTO, GetUser class, and GetUserEndPoint
Added a new `UserDTO` record in the `RSAllies.Api.Contracts` namespace with read-only properties. Implemented a `GetUser` class in the `RSAllies.Api.Features.Users` namespace with a `Query` class and an internal `Handler` class for fetching user data from the database. Also added a `GetUserEndPoint` class that maps a GET request to the `/api/users/{id}` endpoint, returning user data or a 404 status.