`Add CreateUser feature and seal Handler in GetUser`
Implemented the CreateUser feature with a new `CreateUser` class in the `RSAllies.Api.Features.Users` namespace, which includes a `Command` class with properties for user details and a `Handler` class to check for existing users and create new ones. Also added a `CreateUserEndPoint` class that maps a POST request to a method for creating new users. In addition, the `Handler` class in `GetUser.cs` has been made `sealed` to prevent inheritance. A new package reference `Mapster` version `7.4.0` has been added to the `RSAllies.Api.csproj` file.