Update UI styles in MainLayout.razor and app.css
Updated the `MainLayout.razor` and `app.css` files to improve the website's look and feel. In `MainLayout.razor`, the `FluentHeader` tag's style was updated to change the background color to `#EE3026` and the font family to `Bahnschrift`. The text within the `FluentHeader` tag was also changed from `RSAllies` to `Road Safety Allies`. The `FluentBodyContent` tag's style was updated to change the font family to `Bahnschrift`. In `app.css`, the `font-family` property in the `body` selector was changed from `var(--body-font)` to `Bahnschrift`. These changes will enhance the overall user experience by providing a more visually appealing interface.