Implemented booking feature in RSAllies.Api
Added and modified several files to implement a booking feature in the RSAllies.Api application. The `BookingDto.cs` file was modified to add a new namespace and a new public record. A new file `CreateBookingDto.cs` was created with a new namespace and a public record with several properties. The `CreateBooking.cs` and `DeleteBooking.cs` files were created to handle booking creation and deletion. The `GetUserBookings.cs` file was created to retrieve a user's booking and add a new route to the application.