Updated `GetSessionDateRange` and `Venues.razor` for `VenueId` filtering and UI changes
Updated the `GetSessionDateRange` class to include a new `VenueId` property in the `Query` class and modified the `Handle` method to filter sessions based on `VenueId`. Also, the `GetSessionDateRangeEndPoint` class has been updated to accept `VenueId` as a parameter in the `MapGet` method.

In the `Venues.razor` file, the `@rendermode` directive has been changed to `InteractiveAuto`, the `FluentCard` height has been increased, the `if` condition checking `_venues` count has been removed, and the `FluentProgressRing` has been removed. These changes improve the rendering mode, the card size, and the display of venues.