Added new features and DTOs for Venues and ApiClient
In this commit, several significant changes were made. A new page title and header were added to the Venues page in `Venues.razor`. The `RSAllies.Client` namespace was included in `Program.cs`, and Razor components and an HttpClient named `ApiClient` with a base address of "https://apiservice" were added to the services. A project reference to `RSAllies.Api.csproj` was added in `RSAllies.csproj`. A new class `ApiClient` was created in the `RSAllies.Client` namespace in `ApiClient.cs`. New data transfer objects (DTOs) were created in the `RSAllies.Contracts` namespace in several files. Lastly, a new CSS file was created for the Venues page in `Venues.razor.css`, but no styles have been added yet.