Added venue creation functionality and UI enhancements
In this commit, several changes were made to enhance the user interface and add the ability to create new venues. In `ApiClient.cs`, new namespaces were added and the constructor was modified to include an `IDialogService` parameter. A new method `CreateVenueAsync` was also added to handle the creation of new venues.

In `CreateVenue.razor`, new namespaces were added, and `ApiClient` and `IDialogService` were injected. The `EditForm` and `Submit` method were modified to support the new venue creation functionality.

`MainLayout.razor` was updated with the addition of `FluentMessageBarProvider` and `FluentDialogProvider`.

In `Venues.razor`, a `FluentAnchor` was added to provide a link to the create venue page, and conditional rendering was added to display a `FluentProgressRing` if there are no venues, and a `FluentGrid` with the list of venues otherwise.

Finally, a new CSS class was added to `Venues.razor.css` to style the container of the venues list.