Added password authentication and updated entity configurations
In this commit, we've made several changes to enhance user authentication and update entity configurations.

The `AuthenticateDto` class was changed to a record type with added `Phone` and `Password` properties. A nullable `Password` property was added to the `UserDTO` class and a `Password` property with a maximum length of 12 was added to the `User` class.

The `AuthenticateUser` class was changed to an abstract class with added nested `Command` class and an internal sealed `Handler` class for user authentication. A `Password` property was added to the `Command` class in `CreateUser.cs` and the `Handle` method was updated to include the `Password` in the user creation process.

The `MapGet` method in `GetUser.cs` was updated to return the result directly. The initial database migration and the migration for adding a password to the User entity were deleted.

In `AppDbContextModelSnapshot.cs`, several changes were made to the model builder entity configurations, including altering the type of various properties and adding a `Password` property to the `User` entity.

Finally, the versions of `OpenTelemetry.Instrumentation.AspNetCore` and `OpenTelemetry.Instrumentation.Http` were updated in `RSAllies.ServiceDefaults.csproj`.