`Added GetCurrentBooking feature in GetCurrentUserBooking.cs`
This commit introduces a new feature in the `GetCurrentUserBooking.cs` file. The changes include the importation of several namespaces and the creation of a new abstract class `GetCurrentBooking`. This class contains a nested `Query` class and a nested `Handler` class. The `Query` class implements the `IRequest` interface from `MediatR` and is expected to return a `Result` of type `BookingDto`. The `Handler` class implements the `IRequestHandler` interface from `MediatR` and contains an asynchronous `Handle` method. This method queries the `Bookings` DbSet in the `AppDbContext` for a booking that matches the `UserId` in the `Query` object and is not deleted. If no booking is found, it returns a `Result` failure. Otherwise, it returns the `BookingDto` object.