Updated dotnet-tools and routes, added CreateVenueSessions file
Updated the `dotnet-tools.json` file to include a new tool `dotnet-ef` and modified the `rollForward` property for `swashbuckle.aspnetcore.cli` and `dotnet-ef`. The endpoint route for creating a session in `CreateSession.cs` was changed from `api/sessions` to `api/session`.

Added a new file `CreateVenueSessions.cs` with several new classes and methods for handling the creation of new sessions. This includes the `CreateVenueSessions` abstract class, an internal `Handler` class, and a `CreateVenueSessionsEndPoint` class.

Also, added several new namespaces to the `CreateVenueSessions.cs` file.