Updated `appsettings.json` and added `profile.arm.json`
Updated the `AppDbConnection` string in `appsettings.json` to change the database name and password. Added a new file `profile.arm.json`, an Azure Resource Manager (ARM) template, which defines the infrastructure and configuration for the project. The ARM template includes parameters, variables, resource definitions, and a system-assigned managed identity. The app service is configured to use .NET Core and the app service plan is set to the Standard tier.