Added new properties and updated classes for language support
In this commit, several changes were made to support language differentiation in questions. A new nullable boolean property `IsEnglish` was added to the `QuestionDto` class and `Question` class. The `Question` property in `Choice.cs` was made nullable. Formatting changes were made in `GetCurrentUserBooking.cs` for better readability. The `User` property in `Score.cs` and `UserResponse.cs` was made required. A new `Password` property was added to `User.cs`.

In `CreateQuestion.cs` and `CreateSwahiliQuestion.cs`, the `Mapster` namespace was removed, classes were updated with new properties, and the endpoint class was renamed. The `IsEnglish` property was set to `true` in `CreateQuestion.cs` and `false` in `CreateSwahiliQuestion.cs`.

In `GetQuestions.cs` and `GetSwahiliQuestions.cs`, several namespaces were added, classes were made abstract, new classes were added, and the endpoint class was renamed. The `IsEnglish` property was set to `true` in `GetQuestions.cs` and `false` in `GetSwahiliQuestions.cs`.

The `Handler` class in `GetFilteredSessions.cs` and `AuthenticateUser.cs` was updated to handle new properties. The `System` namespace was removed from migration files. A new `IsEnglish` property was added to the `Question` entity in the database model in `AppDbContextModelSnapshot.cs`.