Added new property and updated packages, no functional changes in some files
In this commit, a new property `CreatedAt` of type `DateTime` was added to the `Score` class in `Score.cs` and to the `userScore` object in the `Handle` method of the `Handler` class in `CreateUserScore.cs`. The versions of `OpenTelemetry.Exporter.OpenTelemetryProtocol` and `OpenTelemetry.Extensions.Hosting` were updated from `1.8.0` to `1.8.1` in `RSAllies.ServiceDefaults.csproj`. A new file `GetUserScore.cs` was added, which includes the `GetUserScore` class with a `Query` class and a `Handler` class. The `Handler` class fetches the user score based on the `Id` property of the `Query` class. If no score is found, it returns a failure result. The `GetUserScoreEndPoint` class was also added to define a route to get the user score. No functional changes were made in `AnswerDto.cs`, `GetAnswers.cs`, and `CreateUserScore.cs`, the code was just restructured.